fix(*): reworked to compile with tsc rather than rollup
This gives the option of using project references which greatly cleans up the repository

BREAKING CHANGE: The default target for ts-config and these packages is now CommonJS, which will
work in NodeJS by default and get resolved in Rollup (with @rollup/plugin-commonjs) and Webpack as
well. Furthermore no longer will packages be compiled with importHelpers so tslib is no longer a
hidden required dependency.