fix(querystring): fixed passing "undefined" and "null" to stringify
Previously the entire querystring would pop out of existence. Now the keys are put into the
querystring, but with empty values. This matches how the NodeJS build in "querystring" library
processes undefined/null.

This shouldn't be a breaking change as your backend should be able to
process empty value queryparams, however be sure to test your code after
upgrading!