feat(eslint-config): bumped dependencies to latest versions
This is primarily for eslint-config-react-app as create-react-app v4 has been released which finally
moves create-react-app users to ESLint v7.

BREAKING CHANGE: Providing the additional ESLint packages through dependencies was messing with a
lot of ESLint setups in other projects so instead all dependendencies are now peerDependencies and
need to be installed by you yourself. Please see the README to find out which packages you need to
install.